requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.1
plotly==5.19.0
pandas==2.2.1
numpy==1.26.4
aiohttp==3.9.3
python-dateutil==2.8.2
pytz==2024.1
